Do not send WS closing frame if never connected
If the WebSocket connection was not performed due to external issues, an
exception is thrown in TCPNetworkModule.start().
When this happens, the WebSocketNetworkModule.stop() method still gets
called and it still tries to write the closing frame, even though it never
connected.
This commit skips this operation and just closes the socket instead.